search-api/requirements/dev.txt

# Everything the developer needs in addition to the production requirements
-r prod.txt

coverage==5.0.4
pylint==2.4.4
pylint-flask==0.6
pytest==5.4.1
pytest-cov==2.8.1
pytest-env==0.6.2
pytest-dotenv==0.4.0
pytest-mock==2.0.0
flake8==3.7.9
flake8-blind-except==0.1.1
flake8-debugger==3.2.1
flake8-docstrings==1.5.0
flake8-isort==2.9.0
flake8-quotes==2.1.1
pep8-naming==0.9.1
autopep8==1.5
pydocstyle==5.0.2
FreezeGun==0.3.15
jsonschema==3.2.0
